+++ saleor/checkout/tests/webhooks/test_exclude_shipping_cache.py
@@ -1,35 +1,58 @@
 import json
+import uuid
+from decimal import Decimal
 from unittest import mock
 
 import graphene
+import pytest
+from measurement.measures import Weight
+from prices import Money
 
-from ....webhook.const import CACHE_EXCLUDED_SHIPPING_TIME
+from ....shipping.interface import ShippingMethodData
 from ....webhook.event_types import WebhookEventSyncType
 from ....webhook.transport.utils import generate_cache_key_for_webhook
-from ...base_plugin import ExcludedShippingMethod
+from ...webhooks.exclude_shipping import (
+    CACHE_EXCLUDED_SHIPPING_TIME,
+    excluded_shipping_methods_for_checkout,
+)
+
+
+@pytest.fixture
+def available_shipping_methods():
+    methods = []
+    for i in range(2):
+        methods.append(
+            ShippingMethodData(
+                id=str(i),
+                price=Money(Decimal(10), "usd"),
+                name=uuid.uuid4().hex,
+                maximum_order_weight=Weight(kg=0),
+                minimum_order_weight=Weight(kg=0),
+                maximum_delivery_days=0,
+                minimum_delivery_days=5,
+            )
+        )
+    return methods
 
 
 @mock.patch("saleor.webhook.transport.synchronous.transport.cache.get")
 @mock.patch("saleor.webhook.transport.synchronous.transport.cache.set")
 @mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
 @mock.patch(
-    "saleor.plugins.webhook.plugin."
-    "generate_excluded_shipping_methods_for_checkout_payload"
+    "saleor.checkout.webhooks.exclude_shipping."
+    "_generate_excluded_shipping_methods_for_checkout_payload"
 )
 def test_excluded_shipping_methods_for_checkout_use_cache(
     mocked_payload,
     mocked_webhook,
     mocked_cache_set,
     mocked_cache_get,
-    webhook_plugin,
     checkout_with_items,
-    available_shipping_methods_factory,
-    shipping_app_factory,
+    available_shipping_methods,
+    app_exclude_shipping_for_checkout,
 ):
     # given
-    shipping_app_factory()
     webhook_reason = "Order contains dangerous products."
-    other_reason = "Shipping is not applicable for this order."
 
     mocked_webhook.return_value = {
         "excluded_methods": [
@@ -45,18 +68,12 @@ def test_excluded_shipping_methods_for_checkout_use_cache(
 
     mocked_cache_get.return_value = (payload, [{"id": "1", "reason": webhook_reason}])
 
-    plugin = webhook_plugin()
-    available_shipping_methods = available_shipping_methods_factory(num_methods=2)
-    previous_value = [
-        ExcludedShippingMethod(id="1", reason=other_reason),
-        ExcludedShippingMethod(id="2", reason=other_reason),
-    ]
-
     # when
-    plugin.excluded_shipping_methods_for_checkout(
+    excluded_shipping_methods_for_checkout(
         checkout_with_items,
         available_shipping_methods=available_shipping_methods,
-        previous_value=previous_value,
+        allow_replica=True,
+        requestor=None,
     )
 
     # then
@@ -69,25 +86,22 @@ def test_excluded_shipping_methods_for_checkout_use_cache(
 @mock.patch("saleor.webhook.transport.synchronous.transport.cache.set")
 @mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
 @mock.patch(
-    "saleor.plugins.webhook.plugin."
-    "generate_excluded_shipping_methods_for_checkout_payload"
+    "saleor.checkout.webhooks.exclude_shipping."
+    "_generate_excluded_shipping_methods_for_checkout_payload"
 )
 def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_empty(
     mocked_payload,
     mocked_webhook,
     mocked_cache_set,
     mocked_cache_get,
-    webhook_plugin,
     checkout_with_items,
-    available_shipping_methods_factory,
-    shipping_app_factory,
+    available_shipping_methods,
+    app_exclude_shipping_for_checkout,
 ):
     # given
-    shipping_app = shipping_app_factory()
-    shipping_webhook = shipping_app.webhooks.get()
+    shipping_webhook = app_exclude_shipping_for_checkout.webhooks.get()
 
     webhook_reason = "Order contains dangerous products."
-    other_reason = "Shipping is not applicable for this order."
 
     webhook_response = {
         "excluded_methods": [
@@ -105,18 +119,12 @@ def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_empty(
 
     mocked_cache_get.return_value = None
 
-    plugin = webhook_plugin()
-    available_shipping_methods = available_shipping_methods_factory(num_methods=2)
-    previous_value = [
-        ExcludedShippingMethod(id="1", reason=other_reason),
-        ExcludedShippingMethod(id="2", reason=other_reason),
-    ]
-
     # when
-    plugin.excluded_shipping_methods_for_checkout(
+    excluded_shipping_methods_for_checkout(
         checkout_with_items,
         available_shipping_methods=available_shipping_methods,
-        previous_value=previous_value,
+        allow_replica=True,
+        requestor=None,
     )
 
     # then
@@ -126,7 +134,7 @@ def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_empty(
         payload_dict,
         shipping_webhook.target_url,
         WebhookEventSyncType.CHECKOUT_FILTER_SHIPPING_METHODS,
-        shipping_app.id,
+        app_exclude_shipping_for_checkout.id,
     )
 
     mocked_cache_set.assert_called_once_with(
@@ -140,25 +148,23 @@ def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_empty(
 @mock.patch("saleor.webhook.transport.synchronous.transport.cache.set")
 @mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
 @mock.patch(
-    "saleor.plugins.webhook.plugin."
-    "generate_excluded_shipping_methods_for_checkout_payload"
+    "saleor.checkout.webhooks.exclude_shipping."
+    "_generate_excluded_shipping_methods_for_checkout_payload"
 )
 def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_payload_different(
     mocked_payload,
     mocked_webhook,
     mocked_cache_set,
     mocked_cache_get,
-    webhook_plugin,
     checkout_with_items,
-    available_shipping_methods_factory,
-    shipping_app_factory,
+    available_shipping_methods,
+    app_exclude_shipping_for_checkout,
 ):
     # given
-    shipping_app = shipping_app_factory()
-    shipping_webhook = shipping_app.webhooks.get()
+
+    shipping_webhook = app_exclude_shipping_for_checkout.webhooks.get()
 
     webhook_reason = "Order contains dangerous products."
-    other_reason = "Shipping is not applicable for this order."
 
     webhook_response = {
         "excluded_methods": [
@@ -177,17 +183,12 @@ def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_payload_dif
 
     mocked_cache_get.return_value = None
 
-    plugin = webhook_plugin()
-    available_shipping_methods = available_shipping_methods_factory(num_methods=2)
-    previous_value = [
-        ExcludedShippingMethod(id="1", reason=other_reason),
-        ExcludedShippingMethod(id="2", reason=other_reason),
-    ]
     # when
-    plugin.excluded_shipping_methods_for_checkout(
+    excluded_shipping_methods_for_checkout(
         checkout_with_items,
         available_shipping_methods=available_shipping_methods,
-        previous_value=previous_value,
+        allow_replica=True,
+        requestor=None,
     )
     # then
     assert mocked_webhook.called
@@ -196,7 +197,7 @@ def test_excluded_shipping_methods_for_checkout_stores_in_cache_when_payload_dif
         payload_dict,
         shipping_webhook.target_url,
         WebhookEventSyncType.CHECKOUT_FILTER_SHIPPING_METHODS,
-        shipping_app.id,
+        app_exclude_shipping_for_checkout.id,
     )
     mocked_cache_get.assert_called_once_with(expected_cache_key)
