+++ saleor/graphql/checkout/tests/benchmark/test_checkout_mutations.py
@@ -569,7 +569,7 @@ def test_create_checkout_with_order_promotion(
 
     # when
     user_api_client.ensure_access_token()
-    with django_assert_num_queries(105):
+    with django_assert_num_queries(106):
         response = user_api_client.post_graphql(MUTATION_CHECKOUT_CREATE, variables)
 
     # then
