+++ saleor/webhook/const.py
@@ -1,4 +1,3 @@
-CACHE_EXCLUDED_SHIPPING_TIME = 60 * 3
 WEBHOOK_CACHE_DEFAULT_TTL: int = 5 * 60  # 5 minutes
 SYNC_WEBHOOK_FAILURE_CACHE_TTL: int = 1  # 1 second
 SYNC_WEBHOOK_FAILURE_SENTINEL = (
