+++ saleor/webhook/tests/subscription_webhooks/subscription_queries.py
@@ -3084,23 +3084,6 @@ PAYMENT_LIST_GATEWAYS = """
     """
 
 
-CHECKOUT_FILTER_SHIPPING_METHODS = """
-subscription{
-  event{
-    ...on CheckoutFilterShippingMethods{
-      checkout{
-        id
-      }
-      shippingMethods{
-        name
-        id
-      }
-    }
-  }
-}
-"""
-
-
 SHIPPING_LIST_METHODS_FOR_CHECKOUT = """
 subscription{
   event{
@@ -3111,52 +3094,6 @@ subscription{
       shippingMethods{
         name
         id
-      }
-    }
-  }
-}
-"""
-
-
-CHECKOUT_FILTER_SHIPPING_METHODS_CIRCULAR_SHIPPING_METHODS = """
-subscription{
-  event{
-    ...on CheckoutFilterShippingMethods{
-      checkout{
-        id
-        shippingMethods{
-          id
-        }
-      }
-    }
-  }
-}
-"""
-
-CHECKOUT_FILTER_SHIPPING_METHODS_AVAILABLE_SHIPPING_METHODS = """
-subscription{
-  event{
-    ...on CheckoutFilterShippingMethods{
-      checkout{
-        id
-        availableShippingMethods{
-          id
-        }
-      }
-    }
-  }
-}
-"""
-
-CHECKOUT_FILTER_SHIPPING_METHODS_AVAILABLE_PAYMENT_GATEWAYS = """
-subscription{
-  event{
-    ...on CheckoutFilterShippingMethods{
-      checkout{
-        id
-        availablePaymentGateways{
-          id
-        }
       }
     }
   }
@@ -3254,25 +3191,6 @@ THUMBNAIL_CREATED = """
           id
           objectId
           mediaUrl
-        }
-      }
-    }
-"""
-
-CHECKOUT_SHIPPING_LIST_AND_FILTER = """
-    subscription {
-      event {
-        ... on CheckoutFilterShippingMethods {
-          __typename
-          checkout {
-            id
-          }
-        }
-        ... on ShippingListMethodsForCheckout {
-          __typename
-          checkout {
-            id
-          }
         }
       }
     }
