+++ saleor/webhook/tests/test_webhook_payloads.py
@@ -10,7 +10,6 @@ import graphene
 import pytest
 from django.core.serializers.json import DjangoJSONEncoder
 from freezegun import freeze_time
-from measurement.measures import Weight
 from prices import Money
 
 from ... import __version__
@@ -32,7 +31,6 @@ from ...payment.interface import RefundData, TransactionActionData, TransactionD
 from ...payment.models import TransactionItem
 from ...plugins.manager import get_plugins_manager
 from ...product.models import ProductVariant, ProductVariantChannelListing
-from ...shipping.interface import ShippingMethodData
 from ...warehouse import WarehouseClickAndCollectOption
 from ..payloads import (
     PRODUCT_VARIANT_FIELDS,
@@ -42,7 +40,6 @@ from ..payloads import (
     generate_checkout_payload_for_tax_calculation,
     generate_collection_payload,
     generate_customer_payload,
-    generate_excluded_shipping_methods_for_checkout_payload,
     generate_fulfillment_lines_payload,
     generate_invoice_payload,
     generate_list_gateways_payload,
@@ -2242,37 +2239,6 @@ def test_generate_checkout_payload(
         "meta": generate_meta(requestor_data=generate_requestor(customer_user)),
         "warehouse_address": ANY,
     }
-
-
-def test_generate_excluded_shipping_methods_for_checkout(checkout):
-    shipping_method = ShippingMethodData(
-        id="123",
-        price=Money(Decimal("10.59"), "USD"),
-        name="shipping",
-        maximum_order_weight=Weight(kg=10),
-        minimum_order_weight=Weight(g=1),
-        maximum_delivery_days=10,
-        minimum_delivery_days=2,
-    )
-    response = json.loads(
-        generate_excluded_shipping_methods_for_checkout_payload(
-            checkout, [shipping_method]
-        )
-    )
-
-    assert "checkout" in response
-    assert response["shipping_methods"] == [
-        {
-            "id": graphene.Node.to_global_id("ShippingMethod", "123"),
-            "price": "10.59",
-            "currency": "USD",
-            "name": "shipping",
-            "maximum_order_weight": "10.0:kg",
-            "minimum_order_weight": "1.0:g",
-            "maximum_delivery_days": 10,
-            "minimum_delivery_days": 2,
-        }
-    ]
 
 
 def test_generate_requestor_returns_dict_with_user_id_and_user_type(staff_user, rf):
